Fix height bug for everything in the app container
Took me a while to figure this out. All the re-com examples don't work correctly if the app container is not set to full height. everything will just be the minimum height. I suppose that this should be the default because the re-com documentation seems to implicitly assume that full height works out of the box. Also, body is set to 100% height in the css, too, so I think it makes sense to move the height into the next-higher level too.